requirements.txt
feed-utils==0.6.4
pymongo
apscheduler
requests
flask
flask-classy
flask-cors
kafka-python
docker
psycopg2
bs4
pyjwt
